refactor: 使用 Math.clamp 和现代 Java 特性重构代码
- 将多个 Math.max/min 组合替换为 Math.clamp 以提高可读性
- 使用记录类（record）替换内部类
- 使用方法引用（如 Objects::isNull）简化代码
- 移除未使用的导入语句
- 提取重复逻辑到工具类（如 RagSearchSupport.toNonBlank）
- 简化条件判断和流操作

+++ src/main/java/com/example/EnterpriseRagCommunity/service/ai/AiRerankService.java
@@ -609,7 +609,7 @@ public class AiRerankService {
                 if (idxObj == null) idxObj = readIntLike(n.get("i"));
                 if (idxObj == null) idxObj = readIntLike(n.get("position"));
                 if (idxObj == null) idxObj = readIntLike(n.get("id"));
-                int idx = idxObj == null ? pos : idxObj.intValue();
+                int idx = idxObj == null ? pos : idxObj;
                 Double score = readDoubleLike(n.get("relevance_score"));
                 if (score == null) score = readDoubleLike(n.get("score"));
                 if (score == null) score = readDoubleLike(n.get("relevanceScore"));
